Update Dashboard and Alert headers with the Query one (#4710)
