+++ client/app/pages/alert/components/Title.jsx
@@ -5,11 +5,13 @@ import { getDefaultName } from "../Alert";
 
 import { Alert as AlertType } from "@/components/proptypes";
 
+import "./Title.less";
+
 export default function Title({ alert, editMode, name, onChange, children }) {
   const defaultName = getDefaultName(alert);
   return (
-    <div className="p-b-10 m-l-0 m-r-0 page-header--new">
-      <div className="d-flex">
+    <div className="alert-header">
+      <div className="alert-title">
         <h3>
           {editMode && alert.query ? (
             <Input
@@ -22,8 +24,8 @@ export default function Title({ alert, editMode, name, onChange, children }) {
             name || defaultName
           )}
         </h3>
-        <span className="alert-actions">{children}</span>
       </div>
+      <div className="alert-actions">{children}</div>
     </div>
   );
 }
