+++ client/app/pages/dashboards/DashboardPage.less
@@ -20,92 +20,6 @@
   display: flex;
 }
 
-.dashboard-header {
-  padding: 0 15px !important;
-  margin: 0 0 10px !important;
-  position: -webkit-sticky; // required for Safari
-  position: sticky;
-  background: #f6f7f9;
-  z-index: 99;
-  width: 100%;
-  top: 0;
-
-  @media @mobileBreakpoint {
-    & {
-      padding: 0 !important;
-      position: static;
-    }
-  }
-
-  h3 {
-    margin: 0.2em 0;
-    line-height: 1.3;
-    font-weight: 500;
-  }
-
-  .profile-image {
-    width: 16px;
-    height: 16px;
-    border-radius: 100%;
-    margin: 3px 5px 0 5px;
-  }
-
-  .tags-control a {
-    opacity: 0;
-    transition: opacity 0.2s ease-in-out;
-  }
-
-  &:hover {
-    .tags-control a {
-      opacity: 1;
-    }
-  }
-}
-
-.dashboard-control {
-  margin: 8px 0;
-
-  .icon-button {
-    width: 32px;
-    padding: 0 10px;
-  }
-
-  .save-status {
-    vertical-align: middle;
-    margin-right: 7px;
-    font-size: 12px;
-    text-align: left;
-    display: inline-block;
-
-    &[data-saving] {
-      opacity: 0.6;
-      width: 45px;
-
-      &:after {
-        content: "";
-        animation: saving 2s linear infinite;
-      }
-    }
-
-    &[data-error] {
-      color: #f44336;
-    }
-  }
-}
-
-@keyframes saving {
-  0%,
-  100% {
-    content: ".";
-  }
-  33% {
-    content: "..";
-  }
-  66% {
-    content: "...";
-  }
-}
-
 .add-widget-container {
   background: #fff;
   border-radius: @redash-radius;
