+++ client/app/pages/queries/components/QueryPageHeader.less
@@ -2,7 +2,7 @@
   display: flex;
   flex-wrap: wrap;
   align-items: stretch;
-  margin: -5px 0;
+  margin-bottom: -5px;
 
   & > div {
     padding: 5px 0;
@@ -22,6 +22,10 @@
     .page-title {
       h3 {
         margin: 0 10px 0 0 !important;
+
+        @media (max-width: 767px) {
+          font-size: 18px;
+        }
       }
     }
   }
